backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
beautifulsoup4==4.14.3
requests==2.32.5
selenium==4.41.0
webdriver-manager==4.0.2
supabase==2.28.0
python-dotenv==1.2.1
